Split Plane tests from test_geometry.py
// edited by skirpichev
